parseable maps: deduplicate impl and rename
The implementation of hydrate and reconcile was identical modulo error
handling to the implementations in `hydrate::map` and `reconcile::map`
respectively. Generalize the implementations in `hydrate::map` and
`reconcile::map` and implement `parse_fromstr` in terms of these
generalized implementations.

Then, rename `parse_fromstr` to `map_with_parseable_keys` as this is
more descriptive when reading the incantation in a derive macro and a
better signpost in the documentation index page.
