+++ autosurgeon/src/hydrate.rs
@@ -4,7 +4,7 @@ use std::borrow::Cow;
 use crate::{Prop, ReadDoc};
 
 mod impls;
-mod map;
+pub(crate) mod map;
 
 /// A type which can be hydrated from an automerge document
 ///
