--- autosurgeon/src/parse_fromstr.rs
@@ -1,98 +1,0 @@
-//! Derive macro adaptors for maps with parseable and printable keys.
-//!
-//! The default implementations of [`Reconcile`] and [`Hydrate`] for
-//! [`HashMap`][std::collections::HashMap] and [`BTreeMap`][std::collections::BTreeMap]
-//! require the key to implement [`AsRef<str>`] and [`From<String>`], respectively.  This is not
-//! always possible: for example, a hash map with integer keys cannot be represented that way.
-//! As a solution, this module offers `with`-adaptors for derive macros, which rely on [`ToString`]
-//! and [`FromStr`]:
-//!
-//! ```
-//! # use autosurgeon::{Reconcile, Hydrate};
-//! #[derive(Reconcile, Hydrate)]
-//! struct MyDocument {
-//!     #[autosurgeon(with = "autosurgeon::parse_fromstr")]
-//!     items: std::collections::HashMap<u16, String>,
-//! }
-//! ```
-//!
-//! Note that these adaptors aren't limited to the standard library maps: they work for any
-//! collection implementing [`IntoIterator`] (for [`Reconcile`]) and [`FromIterator`] (for
-//! [`Hydrate`]).
-use std::{error, str::FromStr};
-
-use automerge::{ObjType, Value};
-
-use crate::{
-    reconcile::{LoadKey, MapReconciler},
-    Hydrate, HydrateError, Prop, Reconcile, Reconciler,
-};
-
-pub fn reconcile<'a, K, V, I, R>(items: I, mut reconciler: R) -> Result<(), R::Error>
-where
-    K: ToString + 'a,
-    V: Reconcile + 'a,
-    I: IntoIterator<Item = (&'a K, &'a V)>,
-    R: Reconciler,
-{
-    let mut m = reconciler.map()?;
-    for (k, val) in items.into_iter() {
-        let k = k.to_string();
-        if let LoadKey::Found(new_key) = val.key() {
-            if let LoadKey::Found(existing_key) = m.hydrate_entry_key::<V, _>(&k)? {
-                if existing_key != new_key {
-                    m.replace(k, val)?;
-                    continue;
-                }
-            }
-        }
-        m.put(k, val)?;
-    }
-    Ok(())
-}
-
-pub fn hydrate<'a, D, K, V, M>(
-    doc: &'a D,
-    obj: &automerge::ObjId,
-    prop: Prop<'a>,
-) -> Result<M, crate::HydrateError>
-where
-    D: crate::ReadDoc,
-    K: FromStr,
-    K::Err: error::Error + 'static,
-    V: Hydrate,
-    M: FromIterator<(K, V)>,
-{
-    let obj = match doc.get(obj, &prop)? {
-        Some((Value::Object(ObjType::Map), id)) => id,
-        _ => {
-            return Err(HydrateError::unexpected(
-                "a map",
-                "something else".to_string(),
-            ))
-        }
-    };
-    let Some(obj_type) = doc.object_type(&obj) else {
-        return Err(HydrateError::unexpected("a map", "a scalar value".to_string()))
-    };
-    match obj_type {
-        ObjType::Map | ObjType::Table => doc
-            .map_range(obj.clone(), ..)
-            .map(move |(key, _, _)| {
-                let val = V::hydrate(doc, &obj, key.into())?;
-                let key_parsed: K = key
-                    .parse()
-                    .map_err(|err: K::Err| HydrateError::Parse(err.into()))?;
-                Ok((key_parsed, val))
-            })
-            .collect(),
-        ObjType::Text => Err(HydrateError::unexpected(
-            "a map",
-            "a text object".to_string(),
-        )),
-        ObjType::List => Err(HydrateError::unexpected(
-            "a map",
-            "a list object".to_string(),
-        )),
-    }
-}
